renamed filepicker_rails.rb to filepicker-rails.rb
