+++ lib/filepicker-rails.rb
